requirements.txt
django
djangorestframework
requests  # for the usage example
